upload

+++ src/main/java/ys/project/repository/DocRepo.java
@@ -7,5 +7,5 @@ import ys.project.model.GFDDoc;
  * Created by User on 08.11.2019.
  *  Репозиторий основных документов
  */
-public interface DocRepo extends JpaRepository<GFDDoc, Integer> {
+public interface DocRepo extends JpaRepository<GFDDoc, Long> {
 }
